requirements.txt
asgiref==3.3.1
Django==3.1.7
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
